pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudmigration-benchmark"
version = "0.1.0"
description = "AI Agent评测框架 - 专注于云迁移场景的运维能力评估"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mageric", email = "mageric@example.com"}
]
keywords = ["ai", "benchmark", "cloud-migration", "evaluation", "agent-testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "gradio>=6.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "plotly>=5.18.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
cmb = "cloudmigration_benchmark.web.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cloudmigration_benchmark = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
